+++ examples/nostr-client-example-application/src/main/java/org/tbk/nostr/example/client/MainApplicationRunner.java
@@ -51,11 +51,11 @@ class MainApplicationRunner implements ApplicationRunner, DisposableBean {
                 .delaySubscription(Duration.ofSeconds(5))
                 .subscribe(it -> {
                     String idHex = HexFormat.of().formatHex(it.getId().toByteArray());
-                    String noteId = Nip19.toNote(EventId.fromHex(idHex));
+                    String noteId = Nip19.encodeNote(EventId.fromHex(idHex));
                     String displayId = "%s (%s)".formatted(noteId, idHex);
 
                     String pubkeyHex = HexFormat.of().formatHex(it.getPubkey().toByteArray());
-                    String npub = Nip19.toNpub(MorePublicKeys.fromHex(pubkeyHex));
+                    String npub = Nip19.encodeNpub(MorePublicKeys.fromHex(pubkeyHex));
                     String displayPubkey = "%s (%s)".formatted(npub, pubkeyHex);
                     log.info("""
                             [ALL] New event: