+++ examples/nostr-shell-example-application/src/main/java/org/tbk/nostr/example/shell/command/PersonaCommand.java
@@ -46,8 +46,8 @@ class PersonaCommand {
                 .put("keyPath", account.getPath().asString('\''))
                 .put("privateKey", account.getPrivateKey().toHex())
                 .put("publicKey", account.getPublicKey().value.toHex())
-                .put("nsec", Nip19.toNsec(account.getPrivateKey()))
-                .put("npub", Nip19.toNpub(account.getPublicKey()))
+                .put("nsec", Nip19.encodeNsec(account.getPrivateKey()))
+                .put("npub", Nip19.encodeNpub(account.getPublicKey()))
                 .end()
                 .finish();
     }