+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/Nip19.java
@@ -20,63 +20,63 @@ public final class Nip19 {
         throw new UnsupportedOperationException();
     }
 
-    public static XonlyPublicKey fromNpub(String bech32) {
+    public static XonlyPublicKey decodeNpub(String bech32) {
         return Codecs.decode(bech32, XonlyPublicKey.class);
     }
 
-    public static String toNpub(XonlyPublicKey data) {
+    public static String encodeNpub(XonlyPublicKey data) {
         return Codecs.encode(EntityType.NPUB.getHrp(), data);
     }
 
-    public static PrivateKey fromNsec(String bech32) {
+    public static PrivateKey decodeNsec(String bech32) {
         return Codecs.decode(bech32, PrivateKey.class);
     }
 
-    public static String toNsec(PrivateKey data) {
+    public static String encodeNsec(PrivateKey data) {
         return Codecs.encode(EntityType.NSEC.getHrp(), data);
     }
 
-    public static EventId fromNote(String bech32) {
+    public static EventId decodeNote(String bech32) {
         return Codecs.decode(bech32, EventId.class);
     }
 
-    public static String toNote(EventId data) {
+    public static String encodeNote(EventId data) {
         return Codecs.encode(EntityType.NOTE.getHrp(), data);
     }
 
-    public static Nprofile fromNprofile(String bech32) {
+    public static Nprofile decodeNprofile(String bech32) {
         return Codecs.decode(bech32, Nprofile.class);
     }
 
-    public static String toNprofile(Nprofile data) {
+    public static String encodeNprofile(Nprofile data) {
         return Codecs.encode(EntityType.NPROFILE.getHrp(), data);
     }
 
-    public static String toNprofile(XonlyPublicKey publicKey) {
-        return toNprofile(publicKey, Collections.emptyList());
+    public static String encodeNprofile(XonlyPublicKey publicKey) {
+        return encodeNprofile(publicKey, Collections.emptyList());
     }
 
-    public static String toNprofile(XonlyPublicKey publicKey, Collection<RelayUri> relays) {
-        return toNprofile(Nprofile.builder()
+    public static String encodeNprofile(XonlyPublicKey publicKey, Collection<RelayUri> relays) {
+        return encodeNprofile(Nprofile.builder()
                 .publicKey(publicKey)
                 .relays(relays)
                 .build());
     }
 
-    public static Nevent fromNevent(String bech32) {
+    public static Nevent decodeNevent(String bech32) {
         return Codecs.decode(bech32, Nevent.class);
     }
 
-    public static String toNevent(Nevent data) {
+    public static String encodeNevent(Nevent data) {
         return Codecs.encode(EntityType.NEVENT.getHrp(), data);
     }
 
-    public static String toNevent(Event event) {
-        return toNevent(event, Collections.emptyList());
+    public static String encodeNevent(Event event) {
+        return encodeNevent(event, Collections.emptyList());
     }
 
-    public static String toNevent(Event event, Collection<RelayUri> relays) {
-        return toNevent(Nevent.builder()
+    public static String encodeNevent(Event event, Collection<RelayUri> relays) {
+        return encodeNevent(Nevent.builder()
                 .eventId(EventId.of(event.getId().toByteArray()))
                 .relays(relays)
                 .publicKey(MorePublicKeys.fromEvent(event))
@@ -84,26 +84,26 @@ public final class Nip19 {
                 .build());
     }
 
-    public static Naddr fromNaddr(String bech32) {
+    public static Naddr decodeNaddr(String bech32) {
         return Codecs.decode(bech32, Naddr.class);
     }
 
-    public static String toNaddr(Naddr data) {
+    public static String encodeNaddr(Naddr data) {
         return Codecs.encode(EntityType.NADDR.getHrp(), data);
     }
 
-    public static String toNaddr(EventUri eventUri) {
-        return toNaddr(eventUri, Collections.emptyList());
+    public static String encodeNaddr(EventUri eventUri) {
+        return encodeNaddr(eventUri, Collections.emptyList());
     }
 
-    public static String toNaddr(EventUri eventUri, Collection<RelayUri> relays) {
-        return toNaddr(Naddr.builder()
+    public static String encodeNaddr(EventUri eventUri, Collection<RelayUri> relays) {
+        return encodeNaddr(Naddr.builder()
                 .eventUri(eventUri)
                 .relays(relays)
                 .build());
     }
 
-    public static String toNaddr(Event event) {
+    public static String encodeNaddr(Event event) {
         if (!Nip1.isReplaceableEvent(event) && !Nip1.isAddressableEvent(event)) {
             throw new IllegalArgumentException("Event must be replaceable or addressable: Got kind %d.".formatted(event.getKind()));
         }
@@ -111,7 +111,7 @@ public final class Nip19 {
         XonlyPublicKey publicKey = MorePublicKeys.fromEvent(event);
 
         if (Nip1.isReplaceableEvent(event)) {
-            return toNaddr(EventUri.of(event.getKind(), publicKey.value.toHex()));
+            return encodeNaddr(EventUri.of(event.getKind(), publicKey.value.toHex()));
         }
 
         String dTagValue = MoreTags.findByNameSingle(event, IndexedTag.d)
@@ -119,6 +119,6 @@ public final class Nip19 {
                 .map(it -> it.getValues(0))
                 .orElseThrow(() -> new IllegalStateException("Missing or conflicting '%s' tag.".formatted(IndexedTag.d)));
 
-        return toNaddr(EventUri.of(event.getKind(), publicKey.value.toHex(), dTagValue));
+        return encodeNaddr(EventUri.of(event.getKind(), publicKey.value.toHex(), dTagValue));
     }
 }