+++ nostr/nostr-nip19/src/main/java/org/tbk/nostr/nip19/codec/util/Ints.java
@@ -1,6 +1,6 @@
 package org.tbk.nostr.nip19.codec.util;
 
-public class Ints {
+public final class Ints {
     private Ints() {
         throw new UnsupportedOperationException();
     }